video_core: Restore Frame Skip functionality
This was a feature of citra that got removed on 2016, due to this feature being broken for most of games.

Actually, I managed to fix this by ignoring changed applied to pica_core.cpp about trigger irq (command_processor.cpp for old citra code)

Huge thanks to rtiangha for adapting the old code to the new one.

Co-Authored-By: Reg Tiangha <rtiangha@users.noreply.github.com>